# internal/controller/barang_serial/barang_serial_controller.go
package barang_serial

import (
	"strconv"

	"github.com/gofiber/fiber/v2"

	"github.com/mfaisal-Ash/inventory-backend/internal/middleware"
	barangSerialRepo "github.com/mfaisal-Ash/inventory-backend/internal/repositories/barang_serial"
	"github.com/mfaisal-Ash/inventory-backend/pkg/constant"
	"github.com/mfaisal-Ash/inventory-backend/pkg/utils"
)

const Module = constant.ModuleKelolaBarang

func parseIDParam(c *fiber.Ctx) (uint, error) {
	id, err := strconv.ParseUint(c.Params("id"), 10, 64)
	if err != nil {
		return 0, err
	}
	return uint(id), nil
}

func (h *Controller) List(c *fiber.Ctx) error {
	p := utils.PaginationFromContext(c)
	barangID, _ := strconv.ParseUint(c.Query("barang_id", "0"), 10, 64)
	gudangID, _ := strconv.ParseUint(c.Query("gudang_id", "0"), 10, 64)
	bmItemID, _ := strconv.ParseUint(c.Query("barang_masuk_item_id", "0"), 10, 64)
	bkItemID, _ := strconv.ParseUint(c.Query("barang_keluar_item_id", "0"), 10, 64)
	f := barangSerialRepo.Filter{
		BarangID:           uint(barangID),
		GudangID:           uint(gudangID),
		Status:             c.Query("status", ""),
		BarangMasukItemID:  uint(bmItemID),
		BarangKeluarItemID: uint(bkItemID),
	}
	list, total, err := h.repo.List(p, f)
	if err != nil {
		return utils.Fail(c, fiber.StatusInternalServerError, "gagal mengambil daftar unit barang", nil)
	}
	return utils.OKWithMeta(c, "daftar unit barang berhasil diambil", list, utils.BuildPaginationMeta(p, total))
}

func (h *Controller) Create(c *fiber.Ctx) error {
	var req CreateRequest
	if !utils.ParseAndValidate(c, &req) {
		return nil
	}
	if _, err := h.barangRepo.FindByID(req.BarangID); err != nil {
		return utils.Fail(c, fiber.StatusBadRequest, constant.ErrSerialBarangTidakAda, nil)
	}
	s, err := h.repo.Create(req.BarangID, req.GudangID, req.SerialNumber, req.Catatan)
	if err != nil {
		return utils.Fail(c, fiber.StatusConflict, err.Error(), nil)
	}
	return utils.Created(c, "unit berhasil didaftarkan, stok barang ikut bertambah", s)
}

func (h *Controller) Detail(c *fiber.Ctx) error {
	id, err := parseIDParam(c)
	if err != nil {
		return utils.Fail(c, fiber.StatusBadRequest, "id unit tidak valid", nil)
	}
	s, err := h.repo.FindByID(id)
	if err != nil {
		return utils.Fail(c, fiber.StatusNotFound, "unit tidak ditemukan", nil)
	}
	nomorMasuk, nomorKeluar, err := h.repo.RiwayatDokumen(s)
	if err != nil {
		return utils.Fail(c, fiber.StatusInternalServerError, "gagal mengambil riwayat unit", nil)
	}
	return utils.OK(c, "detail unit berhasil diambil", DetailResponse{
		BarangSerial: s, NomorBarangMasuk: nomorMasuk, NomorBarangKeluar: nomorKeluar,
	})
}

func (h *Controller) Cari(c *fiber.Ctx) error {
	sn := c.Params("sn")
	if sn == "" {
		return utils.Fail(c, fiber.StatusBadRequest, "nomor seri wajib diisi", nil)
	}
	s, err := h.repo.FindBySerial(sn)
	if err != nil {
		return utils.Fail(c, fiber.StatusNotFound, constant.ErrSerialTidakDitemukan, nil)
	}
	nomorMasuk, nomorKeluar, err := h.repo.RiwayatDokumen(s)
	if err != nil {
		return utils.Fail(c, fiber.StatusInternalServerError, "gagal mengambil riwayat unit", nil)
	}
	return utils.OK(c, "unit ditemukan", DetailResponse{
		BarangSerial: s, NomorBarangMasuk: nomorMasuk, NomorBarangKeluar: nomorKeluar,
	})
}

func (h *Controller) Ringkasan(c *fiber.Ctx) error {
	barangID, err := strconv.ParseUint(c.Params("barang_id"), 10, 64)
	if err != nil {
		return utils.Fail(c, fiber.StatusBadRequest, "id barang tidak valid", nil)
	}
	tersedia, terpasang, rusak, err := h.repo.CountByBarang(uint(barangID))
	if err != nil {
		return utils.Fail(c, fiber.StatusInternalServerError, "gagal mengambil ringkasan unit", nil)
	}
	return utils.OK(c, "ringkasan unit berhasil diambil", RingkasanResponse{
		BarangID: uint(barangID), Tersedia: tersedia, Terpasang: terpasang, Rusak: rusak,
	})
}

func (h *Controller) UpdateStatus(c *fiber.Ctx) error {
	id, err := parseIDParam(c)
	if err != nil {
		return utils.Fail(c, fiber.StatusBadRequest, "id unit tidak valid", nil)
	}
	var req UpdateStatusRequest
	if !utils.ParseAndValidate(c, &req) {
		return nil
	}
	s, err := h.repo.UpdateStatusManual(id, req.Status, req.Catatan)
	if err != nil {
		return utils.Fail(c, fiber.StatusConflict, err.Error(), nil)
	}
	return utils.OK(c, "status unit berhasil diperbarui", s)
}

func (h *Controller) Delete(c *fiber.Ctx) error {
	id, err := parseIDParam(c)
	if err != nil {
		return utils.Fail(c, fiber.StatusBadRequest, "id unit tidak valid", nil)
	}
	if err := h.repo.Delete(id); err != nil {
		return utils.Fail(c, fiber.StatusInternalServerError, "gagal menghapus unit", nil)
	}
	return utils.OK(c, "unit berhasil dihapus", nil)
}

func (h *Controller) RegisterRoutes(router fiber.Router) {
	g := router.Group("/barang-serial", middleware.JWTAuth(h.jwtSvc))

	view := middleware.RequirePermission(h.roleRepo, Module, constant.ActionView)
	tambah := middleware.RequirePermission(h.roleRepo, Module, constant.ActionTambah)
	edit := middleware.RequirePermission(h.roleRepo, Module, constant.ActionEdit)
	onlyStaff := middleware.RequireRole(constant.RoleSuperAdmin, constant.RoleAdmin)

	g.Get("/", view, h.List)
	g.Post("/", tambah, h.Create)
	g.Get("/cari/:sn", view, h.Cari)
	g.Get("/ringkasan/:barang_id", view, h.Ringkasan)
	g.Get("/:id", view, h.Detail)
	g.Patch("/:id/status", edit, h.UpdateStatus)
	g.Delete("/:id", onlyStaff, edit, h.Delete)
}